Added container_name to docker compose files
Added container_name to docker compose files to make the initial docker
documentation working again. Because docker exec command is assuming
that the container is named passbolt, but when using docker compose the
container name is service-conrainer-number eg: passbolt-passbolt-1.

With this change the container is named passbolt as stated in the
documentation.